move non operational functions to util namespace
